Fix notebook links url paths
