--- content/predictive_ml_models/Databricks/_index.md
@@ -1,9 +1,0 @@
-+++
-title="Databricks"
-chapter = false
-weight = 2
-pre = ""
-post = ""
-[_build]
-  render = false
-+++
