--- content/predictive_ml_models/Snowflake/_index.md
@@ -1,9 +1,0 @@
-+++
-title="Snowflake"
-chapter = false
-weight = 3
-pre = ""
-post = ""
-[_build]
-  render = false
-+++
