--- content/predictive_ml_models/Snowflake/static/connection.json
@@ -1,9 +1,0 @@
-{
-  "account": "",
-  "user": "",
-  "password": "",
-  "role": "",
-  "warehouse": "",
-  "database": "",
-  "schema": ""
-}
