+++ content/predictive_ml_models/snowflake/model_training.md
@@ -1,0 +1,12 @@
++++
+title = "Model Training"
+weight = 1
++++
+
+#### Train predictive model using LightGBM and MLflow
+
+This propensity to convert solution provides a very flexible way to identify who among your audience is most likely to actually engage with you, for example request a demo / sign up for a trial, purchase a first service/product, request an upgrade, accept an offer etc…
+
+Clone this [notebook](https://github.com/snowplow/composable-cdp-with-predictive-ml-modeling-accelerator/blob/main/content/predictive_ml_models/snowflake/static/snowflake_training_model.ipynb) into your own workspace to run this model for yourself.
+
+{{< jupyter snowflake_training_model 500 >}}
