+++ content/predictive_ml_models/snowflake/static/connection.json
@@ -1,0 +1,9 @@
+{
+  "account": "",
+  "user": "",
+  "password": "",
+  "role": "",
+  "warehouse": "",
+  "database": "",
+  "schema": ""
+}
